fix(vault-import): dedupe duplicate-publicKey keyShares on import
iOS stores one keyShare per chainPublicKey entry, so chains in the same
BIP32 family (e.g. Base / Arbitrum / Ethereum) produce multiple keyShares
sharing a single publicKey. `fromCommVault`'s `chainKeyShares` assignment
loop overwrites on collisions — last write wins — so those chains ended
up with the wrong local share and keysign failed with a `final_session_id`
mismatch against the Vultiserver share.

iOS resolves the same situation via `keyshares.first(where: pubkey == X)`.
Drop duplicate-publicKey keyShares (keep first) before handing off to
`fromCommVault` to match that semantic. Renames the existing helper to
`normalizeImportedCommVault` since it now handles both the master-share
synthesis and the dedup.
